requirements.txt
Django==1.8.3

Celery==3.1.18

django-compressor==1.5

djangorestframework==3.1.3

django-rest-swagger==0.3.4

## Image processing
Pillow==2.5.3

## CORS
django-cors-headers==1.1.0

django-filter==0.10.0

requests==2.7.0

python-dateutil

pytz

geopy==1.10.0

sqlparse==0.1.16
